Remove bga easter egg obfuscation
This cleans up globals with intentially misleading names, or that overwrote other globals created by the game.
In short: having RandomVanish or NoHolds enabled in the songwheel has the chance of swapping the hearts to another image.
Additionally, the code for NoHolds is randomly generated, which made it so a player can randomly trigger the easter egg